refactor(file name)[OK-95] Change file name to undo-redo-command.ts and update imports
