+++ client/src/interfaces/undo-redo-command.ts
@@ -1,0 +1,4 @@
+export interface UndoRedoCommand {
+  redo: () => void;
+  undo: () => void;
+}
